+++ platform/src/generics/CMakeLists.txt
@@ -16,6 +16,7 @@ target_sources(RoughPy_Platform PRIVATE
         comparison_trait.cpp
         conversion_trait.cpp
         conversion_factory.cpp
+        conversion_impl.h
         multiprecision_types.cpp
         number_trait.cpp
         polynomial_types.cpp
