+++ platform/src/generics/multiprecision_types/integer_conversion.h
@@ -15,6 +15,9 @@
 #include "roughpy/generics/type_ptr.h"
 #include "roughpy/generics/conversion_trait.h"
 
+
+#include "generics/conversion_impl.h"
+
 namespace rpy {
 namespace generics::conv {
